Extract the `<StatisticItem />` header
